requirements.txt
# Data Collection
praw==7.7.1
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# NLP
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
vaderSentiment==3.3.2

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.18.0
wordcloud==1.9.2

# API and Web
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
PyYAML==6.0.1
python-dateutil==2.8.2 